laporan 70%
