+++ screen/admin/user/user.js
@@ -1,429 +1,10 @@
-import React, {useState, useEffect} from 'react';
-import {
-  View,
-  Text,
-  StyleSheet,
-  FlatList,
-  TouchableOpacity,
-  ActivityIndicator,
-  Image,
-  Linking,
-  Modal,
-  TextInput,
-  Alert,
-} from 'react-native';
-import {Pressable} from 'react-native';
-import Ionicons from 'react-native-vector-icons/Ionicons';
-import FontAwesome from 'react-native-vector-icons/FontAwesome';
-import {Dropdown} from 'react-native-element-dropdown';
-import axios from 'axios';
+import React from 'react';
+import {View, Text, StyleSheet} from 'react-native';
 
-export default function Presensi() {
-  const [data, setData] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [expandedId, setExpandedId] = useState(null);
-  const [currentPage, setCurrentPage] = useState(1);
-  const [lastPage, setLastPage] = useState(1);  
-  const [isModalVisible, setModalVisible] = useState(false);
-  const [declineReason, setDeclineReason] = useState('');
-  const [selectedUuid, setSelectedUuid] = useState(null);
-  const [searchQuery, setSearchQuery] = useState(''); // State untuk search query
-  const [selectedDisplay, setSelectedDisplay] = useState(null);
-  const [activeButton, setActiveButton] = useState('kontrak');
-
-  useEffect(() => {
-    if (activeButton === 'kontrak') {
-      fetchKontrakData(currentPage); // Panggil fetchKontrakData saat tombol Kontrak aktif
-    }
-  }, [activeButton, currentPage, selectedDisplay]); // Tambahkan currentPage & selectedDisplay untuk memuat data sesuai perubahan
-
-  const fetchRealisasiData = async page => {
-    try {
-      setLoading(true);
-      const response = await axios.post(
-        'http://192.168.61.123:8000/api/v1/lock/indexRealisasi',
-        {page},
-        {
-          headers: {
-            Authorization:
-              'Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xOTIuMTY4LjYxLjEyMzo4MDAwXC9hcGlcL3YxXC9hdXRoXC9yZWZyZXNoIiwiaWF0IjoxNzM0NjY1NTk4LCJleHAiOjE3MzQ2NzEyMjYsIm5iZiI6MTczNDY2NzYyNiwianRpIjoiZXVXUzAzaEZMQlBCc2FTZCIsInN1YiI6MSwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.Obcgx0eW73lkqdw9xyucv_SBgq9L_gNvmGtcRxH5GMU',
-          },
-        },
-      );
-      setData(response.data.data);
-      setCurrentPage(response.data.current_page);
-      setLastPage(response.data.last_page);
-    } catch (error) {
-      console.error('Error fetching data', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const fetchKontrakData = async page => {
-    try {
-      setLoading(true);
-      const response = await axios.post(
-        'http://192.168.61.123:8000/api/v1/lock/indexKontrak',
-        {page},
-        {
-          headers: {
-            Authorization:
-              'Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xOTIuMTY4LjYxLjEyMzo4MDAwXC9hcGlcL3YxXC9hdXRoXC9yZWZyZXNoIiwiaWF0IjoxNzM0NjY1NTk4LCJleHAiOjE3MzQ2NzEyMjYsIm5iZiI6MTczNDY2NzYyNiwianRpIjoiZXVXUzAzaEZMQlBCc2FTZCIsInN1YiI6MSwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.Obcgx0eW73lkqdw9xyucv_SBgq9L_gNvmGtcRxH5GMU',
-          },
-        },
-      );
-      setData(response.data.data);
-      setCurrentPage(response.data.current_page);
-      setLastPage(response.data.last_page);
-    } catch (error) {
-      console.error('Error fetching data', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleApprove = async uuid => {
-    try {
-      Alert.alert('Konfirmasi', 'Apakah Anda yakin ingin mengonfirmasi?', [
-        {text: 'Batal', style: 'cancel'},
-        {
-          text: 'Ya',
-          onPress: async () => {
-            await axios.post(
-              `http://192.168.2.152:8000/perubahan_absensi/${uuid}/change`,
-              {status: '1', revisi: null},
-            );
-            Alert.alert('Berhasil', 'Konfirmasi berhasil.');
-            fetchData(currentPage);
-          },
-        },
-      ]);
-    } catch (error) {
-      Alert.alert('Error', 'Gagal mengonfirmasi.');
-    }
-  };
-
-  const handleDecline = uuid => {
-    setSelectedUuid(uuid);
-    setModalVisible(true);
-  };
-
-  const submitDecline = async () => {
-    try {
-      await axios.post(
-        `http://192.168.2.152:8000/api/v1/perubahan_absensi/${selectedUuid}/change`,
-        {status: '2', revisi: declineReason},
-        {
-          headers: {
-            Authorization:
-              'Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC8xOTIuMTY4LjIuMTUzOjgwMDBcL2FwaVwvdjFcL2F1dGhcL3JlZnJlc2giLCJpYXQiOjE3MzQzOTk0NDQsImV4cCI6MTczNDQxNDc0MiwibmJmIjoxNzM0NDExMTQyLCJqdGkiOiJhS0xXR2w5Y3pkN0pVM1NMIiwic3ViIjoxLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.BXVqz9M9HJ18YZOo7-3uqMpTXHTS5MXTWNH9uu63NcQ',
-            Accept: 'application/json',
-          },
-        },
-      );
-      Alert.alert('Berhasil', 'Penolakan berhasil.');
-      setModalVisible(false);
-      setDeclineReason('');
-      fetchData(currentPage); // Refresh data
-    } catch (error) {
-      Alert.alert('Error', 'Gagal menolak data.');
-    }
-  };
-
-  const display = [
-    {label: '5', value: 1},
-    {label: '10', value: 2},
-    {label: '25', value: 3},
-    {label: '50', value: 4},
-    {label: '100', value: 5},
-  ];
-
-  const toggleExpand = id => {
-    setExpandedId(expandedId === id ? null : id);
-  };
-
-  const getStatusStyle = status => {
-    switch (status?.toUpperCase()) {
-      case 'DISETUJUI':
-        return styles.approvedStatus;
-      case 'DITOLAK':
-        return styles.rejectedStatus;
-      case 'MENUNGGU':
-        return styles.pendingStatus;
-      default:
-        return styles.defaultStatus;
-    }
-  };
-
-  const handlePress = buttonName => {
-    setActiveButton(buttonName); // Atur tombol aktif
-    if (buttonName === 'kontrak') {
-      fetchKontrakData(currentPage); // Langsung fetch data saat tombol Kontrak dipilih
-    } else {
-      fetchRealisasiData(currentPage); // Langsung fetch data saat tombol Realisasi dipilih      
-    }
-  };
-
-  const TableHeader = () => (
-    <View>
-      <View style={styles.tambahContainer}>
-        <TouchableOpacity style={styles.tambahButton}>
-          <FontAwesome name="plus" size={20} color="#fff" style={styles.icon} />
-          <Text style={styles.tambahText}>TAMBAH</Text>
-        </TouchableOpacity>
-        <View style={styles.kontrakContainer}>
-          <Pressable
-            style={({pressed}) => [
-              styles.button,
-              pressed && styles.buttonPressed,
-              activeButton === 'kontrak' && styles.buttonActive,
-            ]}
-            onPress={() => handlePress('kontrak')}>
-            <View style={styles.kontrakButton}>
-              <FontAwesome
-                name="tint"
-                size={24}
-                color={activeButton === 'kontrak' ? '#A463FC' : 'gray'}
-                style={styles.icon}
-              />
-              <Text
-                style={[
-                  styles.buttonText,
-                  activeButton === 'kontrak' && styles.textActive,
-                ]}>
-                Kontrak
-              </Text>
-            </View>
-          </Pressable>
-
-          <Pressable
-            style={({pressed}) => [
-              styles.button,
-              pressed && styles.buttonPressed,
-              activeButton === 'realisasi' && styles.buttonActive,
-            ]}
-            onPress={() => handlePress('realisasi')}>
-            <View style={styles.kontrakButton}>
-              <FontAwesome
-                name="tint"
-                size={24}
-                color={activeButton === 'realisasi' ? '#A463FC' : 'gray'}
-                style={styles.icon}
-              />
-              <Text
-                style={[
-                  styles.buttonText,
-                  activeButton === 'realisasi' && styles.textActive,
-                ]}>
-                Realisasi
-              </Text>
-            </View>
-          </Pressable>
-        </View>
-      </View>
-      <View style={styles.filterContainer}>
-        <View style={styles.displayContainer}>
-          <Text style={styles.displayText}>Display</Text>
-          <Dropdown
-            style={styles.dropdown}
-            data={display}
-            labelField="label"
-            valueField="value"
-            placeholder="10"
-            value={selectedDisplay}
-            onChange={item => setSelectedDisplay(item.value)}
-            renderItem={item => (
-              <Text style={[styles.dropdownItem, styles.customFont]}>
-                {item.label}
-              </Text>
-            )}
-          />
-        </View>
-        {/* Search Bar */}
-        <View style={styles.searchContainer}>
-          <TextInput
-            style={styles.searchBar}
-            placeholder="Search"
-            value={searchQuery}
-            onChangeText={setSearchQuery}
-          />
-        </View>
-      </View>
-      <View style={styles.tableHeader}>
-        <Text style={[styles.headerCell, styles.numberCell]}>No</Text>
-        <Text style={[styles.headerCell, styles.nameCell]}>Jenis</Text>
-        <Text style={[styles.headerCell, styles.tableStatusCell]}>Status</Text>
-        <View style={styles.expandIconCell} />
-      </View>
-    </View>
-  );
-
-  const renderItem = ({item, index}) => {
-    const isExpanded = expandedId === item.id;
-
-    return (
-      <View style={styles.tableRow}>
-        <TouchableOpacity
-          style={styles.rowHeader}
-          onPress={() => toggleExpand(item.id)}>
-          <Text style={[styles.tableCell, styles.numberCell]}>{index + 1}</Text>
-          <Text
-            style={[styles.tableCell, styles.nameCell]}
-            numberOfLines={1}
-            ellipsizeMode="tail">
-            {item.jenis || '-'}
-          </Text>
-          <View style={styles.statusCellContainer}>
-            <Text
-              style={[
-                styles.tableCell,
-                styles.statusCell,
-                getStatusStyle(item.status),
-              ]}>
-              {item.status || '-'}
-            </Text>
-          </View>
-          <View style={styles.expandIconCell}>
-            <Ionicons
-              name={isExpanded ? 'chevron-up' : 'chevron-down'}
-              size={20}
-              color="#333"
-            />
-          </View>
-        </TouchableOpacity>
-        {isExpanded && (
-          <View style={styles.expandedContent}>
-            <Text style={styles.expandedText}>
-              Tanggal: {item.tanggal || '-'}
-            </Text>
-            <Text style={styles.expandedText}>
-              Jam Masuk: {item.jam_masuk || '-'}
-            </Text>
-            <Text style={styles.expandedText}>
-              Jam Keluar: {item.jam_keluar || '-'}
-            </Text>
-            <View style={styles.filetext}>
-              <Text>File:</Text>
-              <Text
-                style={styles.expandedLinkText}
-                onPress={() => Linking.openURL(item.file)}>
-                File Absensi
-              </Text>
-            </View>
-            <View style={styles.actionContainer}>
-              <TouchableOpacity
-                style={styles.approveButton}
-                onPress={() => handleApprove(item.uuid)}>
-                <Ionicons name="checkmark" size={20} color="white" />
-              </TouchableOpacity>
-              <TouchableOpacity
-                style={styles.declineButton}
-                onPress={() => handleDecline(item.uuid)}>
-                <Ionicons name="close" size={20} color="white" />
-              </TouchableOpacity>
-            </View>
-          </View>
-        )}
-      </View>
-    );
-  };
-
+export default function Lock() {
   return (
     <View style={styles.container}>
-      {/* Header */}
-      <View style={styles.header}>
-        <View style={styles.headerLeft}>
-          <Image
-            source={require('../assets/images/logo.png')}
-            style={styles.logo}
-          />
-        </View>
-        <View style={styles.headerRight}>
-          <TouchableOpacity style={styles.iconWrapper}></TouchableOpacity>
-          <TouchableOpacity style={styles.iconWrapper}>
-            <Ionicons name="person-circle-outline" size={24} color="#333" />
-          </TouchableOpacity>
-        </View>
-      </View>
-      {/* Loading Indicator */}
-      {loading ? (
-        <ActivityIndicator size="large" color="#0000ff" />
-      ) : (
-        <FlatList
-          ListHeaderComponent={TableHeader}
-          data={data}
-          renderItem={renderItem}
-          keyExtractor={item => item.id.toString()}
-          contentContainerStyle={styles.card}
-          ListFooterComponent={
-            <View>
-              <Text style={styles.pageInfo}>
-                Showing page {currentPage} of {lastPage}
-              </Text>
-              <View style={styles.paginationContainer}>
-                <View style={styles.paginationButtons}>
-                  <TouchableOpacity
-                    style={[
-                      styles.pageButton,
-                      currentPage === 1 && styles.disabledButton,
-                    ]}
-                    disabled={currentPage === 1}
-                    onPress={() =>
-                      setCurrentPage(prev => Math.max(prev - 1, 1))
-                    }>
-                    <Text style={styles.pageButtonText}>Previous</Text>
-                  </TouchableOpacity>
-                  <TouchableOpacity
-                    style={[
-                      styles.pageButton,
-                      currentPage === lastPage && styles.disabledButton,
-                    ]}
-                    disabled={currentPage === lastPage}
-                    onPress={() =>
-                      setCurrentPage(prev => Math.min(prev + 1, lastPage))
-                    }>
-                    <Text style={styles.pageButtonText}>Next</Text>
-                  </TouchableOpacity>
-                </View>
-              </View>
-            </View>
-          }
-        />
-      )}
-
-      {/* Modal Input Alasan Penolakan */}
-      <Modal
-        visible={isModalVisible}
-        animationType="fade"
-        transparent
-        onRequestClose={() => setModalVisible(false)}>
-        <View style={styles.modalContainer}>
-          <View style={styles.modalContent}>
-            <Text style={styles.modalTitle}>Tolak Data</Text>
-            <Text style={styles.modalLabel}>Alasan Penolakan:</Text>
-            <TextInput
-              style={styles.modalInput}
-              placeholder="Masukkan alasan"
-              multiline
-              value={declineReason}
-              onChangeText={setDeclineReason}
-            />
-            <View style={styles.modalButtons}>
-              <TouchableOpacity
-                style={styles.cancelButton}
-                onPress={() => setModalVisible(false)}>
-                <Text style={styles.buttonText}>Batal</Text>
-              </TouchableOpacity>
-              <TouchableOpacity
-                style={styles.submitButton}
-                onPress={submitDecline}>
-                <Text style={styles.buttonText}>Tolak</Text>
-              </TouchableOpacity>
-            </View>
-          </View>
-        </View>
-      </Modal>
+      <Text style={styles.text}>Lock Screen</Text>
     </View>
   );
 }
@@ -431,339 +12,10 @@ export default function Presensi() {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#F7F8FB',
-  },
-  card: {
-    backgroundColor: '#FFFFFF',
-    borderRadius: 10,
-    padding: 15,
-    margin: 12,
-    elevation: 3,
-    shadowColor: '#000',
-    shadowOffset: {width: 0, height: 2},
-    shadowOpacity: 0.2,
-    shadowRadius: 4,
-  },
-  tableHeader: {
-    flexDirection: 'row',
-    backgroundColor: '#E0E0E0',
-    paddingVertical: 10,
-    paddingHorizontal: 15,
-    borderTopLeftRadius: 10,
-    borderTopRightRadius: 10,
-  },
-  headerCell: {
-    fontFamily: 'Poppins-SemiBold',
-    fontSize: 13,
-    color: '#333',
-  },
-  tableRow: {
-    backgroundColor: '#FFFFFF',
-    overflow: 'hidden',
-    elevation: 0,
-  },
-  rowHeader: {
-    flexDirection: 'row',
-    padding: 15,
-    backgroundColor: '#F0F0F0',
-    alignItems: 'center',
-  },
-  tableCell: {
-    fontFamily: 'Poppins-Regular',
-    flexWrap: 'wrap',
-    fontSize: 14,
-  },
-  tableStatusCell: {
-    textAlign: 'center',
-    flex: 1,
-    paddingLeft: 0,
-  },
-  numberCell: {
-    width: 50,
-  },
-  nameCell: {
-    flex: 1,
-    overflow: 'hidden',
-  },
-  statusCellContainer: {
-    width: 100,
-  },
-  statusCell: {
-    textAlign: 'center',
-    fontWeight: 'bold',
-  },
-  expandIconCell: {
-    width: 40,
-    alignItems: 'flex-end',
-  },
-  approvedStatus: {
-    color: '#4CAF50',
-  },
-  rejectedStatus: {
-    color: '#F44336',
-  },
-  pendingStatus: {
-    color: '#FFC107',
-  },
-  defaultStatus: {
-    color: '#9E9E9E',
-  },
-  expandedContent: {
-    padding: 15,
-    backgroundColor: '#FAFAFA',
-  },
-  expandedText: {
-    marginBottom: 5,
-    fontSize: 14,
-  },
-  expandedLinkText: {
-    color: 'blue',
-    marginBottom: 5,
-    fontSize: 14,
-  },
-  filetext: {
-    flexDirection: 'row',
-  },
-  actionContainer: {
-    flexDirection: 'row',
-    marginTop: 10,
-    justifyContent: 'flex-end',
-    alignItems: 'center',
-    gap: 10,
-  },
-  actionButton: {
-    padding: 8,
-    backgroundColor: '#F0F0F0',
-    borderRadius: 5,
-  },
-  approveButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '#4CAF50',
-    paddingVertical: 10,
-    paddingHorizontal: 10,
-    borderRadius: 10,
-    elevation: 3,
-    shadowColor: '#000',
-    shadowOffset: {width: 0, height: 2},
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-  },
-  declineButton: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    backgroundColor: '#F44336',
-    paddingVertical: 10,
-    paddingHorizontal: 10,
-    borderRadius: 10,
-    elevation: 3,
-    shadowColor: '#000',
-    shadowOffset: {width: 0, height: 2},
-    shadowOpacity: 0.2,
-    shadowRadius: 2,
-  },
-  paginationButtons: {
-    flexDirection: 'row',
-  },
-  pageButton: {
-    padding: 10,
-    backgroundColor: '#007bff',
-    borderRadius: 5,
-    marginHorizontal: 5,
-  },
-  pageButtonText: {
-    fontFamily: 'Poppins-Regular',
-    fontSize: 13,
-    color: '#fff',
-  },
-  disabledButton: {
-    backgroundColor: '#CCCCCC',
-  },
-  paginationText: {
-    color: 'white',
-    fontWeight: 'bold',
-  },
-  pageInfo: {
-    fontFamily: 'Poppins-Regular',
-    fontSize: 13,
-  },
-  paginationContainer: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-    marginTop: 10,
-  },
-  header: {
-    backgroundColor: '#ffffff',
-    paddingHorizontal: 16,
-    paddingVertical: 12,
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'space-between',
-    elevation: 4,
-    borderBottomLeftRadius: 20,
-    borderBottomRightRadius: 20,
-  },
-  headerLeft: {
-    flex: 1,
-  },
-  logo: {
-    width: 140,
-    height: 40,
-    resizeMode: 'contain',
-  },
-  headerRight: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    justifyContent: 'flex-end',
-    flex: 1,
-  },
-  iconWrapper: {
-    marginLeft: 12,
-  },
-  modalContainer: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-  },
-  modalContent: {
-    width: '80%',
-    backgroundColor: '#fff',
-    padding: 20,
-    borderRadius: 10,
-    elevation: 5,
-  },
-  modalTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    marginBottom: 10,
-  },
-  modalLabel: {
-    fontSize: 14,
-    marginBottom: 5,
-  },
-  modalInput: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    borderRadius: 5,
-    padding: 10,
-    minHeight: 80,
-    marginBottom: 15,
-    textAlignVertical: 'top',
-  },
-  modalButtons: {
-    flexDirection: 'row',
-    justifyContent: 'flex-end',
-  },
-  cancelButton: {
-    backgroundColor: '#ccc',
-    padding: 10,
-    borderRadius: 5,
-    marginRight: 10,
-  },
-  submitButton: {
-    backgroundColor: '#F44336',
-    padding: 10,
-    borderRadius: 5,
-  },
-  searchContainer: {
-    width: 150,
-    backgroundColor: '#FFFFFF',
-  },
-  searchBar: {
-    height: 40,
-    borderColor: '#CCCCCC',
-    borderWidth: 1,
-    borderRadius: 5,
-    paddingHorizontal: 12,
-  },
-  filterContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 15,
-  },
-  displayContainer: {
-    flexDirection: 'row',
-    justifyContent: 'center',
-    alignItems: 'center',
-    marginRight: 20,
-  },
-  displayText: {
-    fontFamily: 'Poppins-Regular',
-    fontSize: 13,
-    marginRight: 8,
-    textAlign: 'center',
-    color: '#3f4254',
-  },
-  dropdown: {
-    height: 40,
-    borderColor: '#CCCCCC',
-    borderWidth: 1,
-    borderRadius: 5,
-    paddingHorizontal: 10,
-    width: 75,
     justifyContent: 'center',
     alignItems: 'center',
   },
-  dropdownItem: {
-    padding: 10,
-    fontSize: 13,
-    color: '#333',
-  },
-  customFont: {
-    fontFamily: 'Poppins-Regular',
-  },
-  tambahContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginBottom: 10,
-  },
-  tambahButton: {
-    flexDirection: 'row',
-    backgroundColor: '#3699FF',
-    width: 90,
-    height: 40,
-    borderRadius: 5,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  tambahText: {
-    marginTop: 1,
-    fontFamily: 'Poppins-Regular',
-    color: 'white',
-    marginLeft: 5,
-    lineHeight: 20,
-    fontSize: 13,
-    textAlignVertical: 'center',
-  },
-  button: {
-    height: 40,
-    width: 90,
-    backgroundColor: '#fff',
-    borderRadius: 5,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  buttonPressed: {
-    backgroundColor: '#fff',
-  },
-  buttonActive: {
-    backgroundColor: '#E6E7F0',
-  },
-  buttonText: {
-    fontFamily: 'Poppins-Regular',
-    fontSize: 14,
-    color: 'grey',
-  },
-  textActive: {
-    fontFamily: 'Poppins-Regular',
-    color: '#A463FC',
-  },
-  kontrakContainer: {
-    flexDirection: 'row',
-  },
-  kontrakButton: {
-    flexDirection: 'row',
-    gap: 5,
+  text: {
+    fontSize: 24,
   },
 });
